Extract a spatial reference identifier

The geovctrs package works with spatial reference identifiers instead of actual CRS objects so that the heavy lifting can be implemented in other packages. However, most geometry formats store an integer spatial reference identifier (SRID) with each feature to propogate this information through calculations. Note that an SRID of 0 is interpreted as "not set".

geo_srid(x)

# S3 method for default
geo_srid(x)

set_geo_srid(x, srid)

as_geo_srid(x)

Arguments

x
A geometry-like object, or one that can be coerced to a geometry-like object using as_geovctr().

srid
A spatial reference identifier, coerced to an integer by as_geo_srid(). These identifiers can and should be managed outside of geovctrs except for 0, which is interpreted as "not set".

Value

An integer vector (one SRID per feature).

Examples

# two points with an SRID
geometries <- c(
  geo_point(geo_xy(259473, 4876249), srid = 26920),
  geo_point(geo_xy(-66, 44), srid = 4326)
)

geo_srid(geometries)

#> [1] 26920  4326

geo_srid(set_geo_srid(geometries, NA))

#> [1] NA NA

# SRIDs are propogated through conversions,
# or discarded with a warning
geo_srid(as_geo_wkb(geometries))

#> [1] 26920  4326

geo_srid(as_geo_xy(geometries))

#> Warning: Dropping SRID in cast to geo_xy()

#> Warning: Dropping SRID in cast to geo_xy()

#> [1] 0 0