Create and validate well-known binary

Like other geo types, geo_wkb() doesn't convert its input but does validate it using validate_geovctrs_wkb(). To skip validation, use new_geovctrs_wkb() with the result of vec_cast(list(...), .ptype = raw()).

geo_wkb(x = list())

parse_wkb(x)

as_geo_wkb(x, ..., include_srid = NA, dimensions = 3, endian = NA)

Arguments

x
A list() of raw() objects, each of which represent well-known binary

...
Unused

include_srid
Use TRUE to always include, FALSE to never include, or NA to only include if the SRID is non-zero.

dimensions
The maximum number of dimensions to consider when generating text.

endian
Use 0 for big endian, 1 for little endian or NA to use the default on your system.

Value

A new_geovctrs_wkb()

Examples

# POINT (30 10) in WKB
wkb_item <- as.raw(
  c(
    0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x3e, 0x40, 0x00, 0x00, 0x00,
    0x00, 0x00, 0x00, 0x24, 0x40
  )
)
wkb <- geo_wkb(list(wkb_item))
wkb

#> <geovctrs_wkb[1]>
#> [1] <POINT (30 10)>

# use as_geo_wkb() to use conversion options
as_geo_wkb(wkb, endian = 0)

#> <geovctrs_wkb[1]>
#> [1] <POINT (30 10)>

# use parse_wkb() to identify parse errors
wkb_item[2] <- as.raw(0x11)
parse_wkb(list(wkb_item))

#> Warning: 1 parsing failure.
#> row col expected     actual                                                                             error                                                      
#> 1   --  geovctrs_wkb 01, 11, 00, 00, 00, 00, 00, 00, 00, 00, 00, 3e, 40, 00, 00, 00, 00, 00, 00, 24, 40 Unrecognized geometry type in WKBReader::readGeometry(): 17
#> See attr(result, 'problems') for more details.

#> <geovctrs_wkb[1]>
#> [1] <NA_wkt_>